+++ src/Pyro4/__init__.py
@@ -60,7 +60,7 @@ config = Configuration()
 del Configuration
 
 # import the required Pyro symbols into this package
-from Pyro4.core import URI, Proxy, Daemon, callback, batch, async
+from Pyro4.core import URI, Proxy, Daemon, callback, batch, async, oneway
 from Pyro4.naming import locateNS, resolve
 from Pyro4.futures import Future
 from Pyro4.constants import VERSION as __version__